refactor: Campaign.Server를 ToolHandler, Summarizer, Combat 모듈로 분리
God Object였던 server.ex(1361줄)를 4개 모듈로 분리:
- server.ex (218줄): GenServer 콜백과 탐험 모드 조율만 담당
- tool_handler.ex (525줄): AI 도구 결과 → 상태 변환 (캐릭터, NPC, 퀘스트, 전투, 레벨업)
- summarizer.ex (356줄): AI 요약 생성 (세션, 컨텍스트, 전투 히스토리, 전투 종료)
- combat.ex (295줄): 전투 흐름 관리 (플레이어 턴, 적 그룹 턴, 라운드 정리)

Combat 모듈에 tool_context 명시적 전달 패턴 도입.
